+++ src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Router from "./Router";
+import Router from "Router";
 
 export default function App() {
   return (
