+++ src/components/commons/Button.js
@@ -1,0 +1,6 @@
+import styles from "styles/Commons.module.scss";
+import React from "react";
+
+export default function Button({ text }) {
+  return <div className={styles.button}>{text}</div>;
+}
