--- src/components/templates/BestProduct.module.scss
@@ -1,5 +1,0 @@
-@import "../../styles/variables";
-
-.layout {
-  padding: 0 20px;
-}
