+++ src/features/hooks/useMobileDetect.js
