+++ src/pages/supply/privacy.js
