+++ src/styles/index.scss
@@ -1,2 +1,2 @@
-@import url(./reset.scss);
-@import url(./variables.scss);
+@import url(reset.scss);
+@import url(variables.scss);
