+++ src/styles/items.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/variables";
+@import "variables";
 
 .container {
   color: var(--cool-gray800);
