+++ src/styles/layout.module.scss
@@ -1,0 +1,39 @@
+@import "variables";
+
+/* GNB */
+
+.navbar {
+  display: flex;
+  align-items: center;
+  padding: 10px 16px;
+  border: 1px solid #dfdfdf;
+
+  @media (min-width: $min-tablet-width) {
+    padding: 10px 34px;
+  }
+
+  @media (min-width: $min-desktop-width) {
+    padding: 10px 200px;
+  }
+}
+
+.menus {
+  flex-grow: 1;
+  font-weight: 700;
+  font-size: 16px;
+  margin-left: 15px;
+  color: var(--cool-gray600);
+  display: flex;
+  align-items: center;
+  gap: 8px;
+
+  @media (min-width: $min-tablet-width) {
+    font-size: 18px;
+    gap: 38px;
+    margin-left: 32px;
+  }
+
+  .focus {
+    color: var(--blue);
+  }
+}
